+++ Magic/CHANGELOG.md
@@ -20,6 +20,7 @@
  - Added `cooldown`, `biomes` and `not_biomes` options to block populators
  - The ColorItem action can be used to color a potion, firework charge or a firework (will change all effects in a firework)
  - Add mgive shortcut for `firework_charge:#FF00AA` to give colored firework charges
+ - Add `infinity_cost_free` wand property, will make the wand cost-free if it has the Infinity enchantment
 
 ## Changes